pyproject.toml
[tool.poetry]
name = "prompt-tester"
version = "0.1.0"
description = ""
authors = ["tankh99 <khanghou99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytelegrambotapi = "^4.21.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
